More physics experiments
Tried using some hinges for the position-based approach with stick constraints (for elbows and knees) to try and have a more local frame rather than relying on neighbor joints.

For the rigid body approach, managed to get closer to getting angular impulses to work on the constraints so that the bodies actually move when they fall, but it's still unbelievably unstable.

I really should've forked this lol