Rework constraints a bit, specify base and joint axis in the puppet
Some joints (ex. arms) are not in T pose where the joint and cone axis are aligned, so we need to be able to specify the axis that the joint is pointing for constraining

Hinge constraint angle restriction was wrong, it was applying to the swing. We need to either account for twist properly (which seems to be annoying), or also add a cone constraint to prevent rotation beyond a certain point.