Refactored shader importing
Collect all the loaded imports and add them to the source list instead of concatenating them to the string in-place. That way we don't need to allocate a huge buffer for the source.

Any comment found on an import line will also cause the entire line to be skipped, makes it easier to handle cases like //import "blah"