Sprite stuff
I don't love this so I'm submitting as-is and shelving the topic for later. Would be better to move forward with a sprite thing type akin to Indy.

There's some bugs here such as the transform not always working (ex. the muzzle flashes). Will probably just revert the change temporarily and revisit later.